#### Details

Issue: #117 

- [x] tl;dr - Improves how we reassemble InfluxDB fields to span annotations:
  - Updates `func annotationsFromRow(...)` to include common logic that we use to reassemble data back from InfluxDB fields to span `annotations`. Now it uses `func annotationsFromEvents(..)` which will discard those annotations that should no be included because were not explicitly save by `Collect(...)` but that are present because `InfluxDB` will return all the fields present on `spans` measurement:
    - Eg. A field named `Client.Request.Method` might contain values for sub-spans but not for root-spans. Since span annotations are non-deterministic, current strategy for query those annotations(fields) is to do `select * from spans ....`, therefore: `Client.Request.Method` will be included in the query result even for root-spans:
![screen shot 2016-03-17 at 14 18 14](https://cloud.githubusercontent.com/assets/1000404/13858540/1eb5c054-ec4d-11e5-8d08-cc9715b0c54c.png)