@chris-ramon 2016-03-18
tl:dr —There's not performance impact introduced to the `InfluxDBStore.Trace(...) & InfluxDBStre.Traces(...)` with this PR.

I went ahead and re-run benchmarks from #114 for `InfluxDBStore.Trace(...)` & `InfluxDBStore.Traces(...)` in order to confirm if there are big performance impact with [changes](https://github.com/sourcegraph/appdash/pull/121/files#diff-b7da646763ca1f5622f8d108c536ce86R422) introduced in this PR because unmarshalling/marshalling via: [`UnmarshalEvents`](https://github.com/sourcegraph/appdash/blob/955366b27d4851653681b18179a6b905f932f2ed/event.go#L134) & [`MarshalEvent`](https://github.com/sourcegraph/appdash/blob/955366b27d4851653681b18179a6b905f932f2ed/event.go#L42).

- `InfluxDBStore.Trace(...)`

  - Before:

```bash
$ go test -run=XYZ -bench BenchmarkInfluxDBStoreTrace. -benchtime=40s > trace
$ more trace

PASS
BenchmarkInfluxDBStoreTrace100-4             200         411849069 ns/op
BenchmarkInfluxDBStoreTrace250-4              50        1051788921 ns/op
BenchmarkInfluxDBStoreTrace1000-4             10        4227713513 ns/op
ok      sourcegraph.com/sourcegraph/appdash     380.156s
```

  - After:

```bash
$ go test -run=XYZ -bench BenchmarkInfluxDBStoreTrace[0-9]+ -benchtime=40s > trace
$ more trace

PASS
BenchmarkInfluxDBStoreTrace100-4             100         412541990 ns/op
BenchmarkInfluxDBStoreTrace250-4              50        1079761891 ns/op
BenchmarkInfluxDBStoreTrace1000-4             10        4208731896 ns/op
ok      sourcegraph.com/sourcegraph/appdash     294.010s
```

- `InfluxDBStore.Traces(...)`

Before:

```bash
$ go test -run=XYZ -bench BenchmarkInfluxDBStoreTracesDefaultPerPage > traces
$ more traces

PASS
BenchmarkInfluxDBStoreTracesDefaultPerPage-4         100          13433622 ns/op
ok      sourcegraph.com/sourcegraph/appdash     10.595s
```

After:

```bash
$ go test -run=XYZ -bench BenchmarkInfluxDBStoreTracesDefaultPerPage > traces
$ more traces
PASS
BenchmarkInfluxDBStoreTracesDefaultPerPage-4         100          13300171 ns/op
ok      sourcegraph.com/sourcegraph/appdash     11.540s
```