@slimsag 2016-03-29
I hope to test this change with our app soon to confirm, but this looks like a very good change. This seems much more correct than #119 and using using MarshalEvent/UnmarshalEvent seems correct to me (but also highlights why Appdash should eventually move to a more simplified storage modal).

I'll follow-up with more issues if I spot them in my testing. Thanks Chris! :+1: